### 2018-07-03 #804 @whedon
[PRE REVIEW]: mzEasy
**Submitting author:** @chasemc (<a href="http://orcid.org/0000-0001-6439-9397">Chase M Clark</a>)
**Repository:** <a href="https://github.com/chasemc/mzEasy" target ="_blank">https://github.com/chasemc/mzEasy</a>
**Version:** 0.1.0
