### 2018-07-03 #804 @chasemc
I did some searching but am currently unsure the reason the PDF isn't compiling.

@biswapriyamisra may be suited for this review
