### 2018-07-07 #804 @biswapriyamisra
1. Fairly quick and easy installation when MSConvert and R 3.5 or so are in place!
2. Useful too for the conversions.
3. More comments and suggestions in the annotated PDF.
4. Useful tutorial and walk through images.

[mzEasy_JOSS_Reviewer_Comments.pdf](https://github.com/openjournals/joss-reviews/files/2172385/mzEasy_JOSS_Reviewer_Comments.pdf)

Thanks for the opportunity to review and all the best. Shall keep adding requests to more functionalities or issues when faced.

Thanks,
Biswa
