### 2018-07-08 #804 @bovee
Sorry, I'm currently on vacation without a computer (it'd be at least a
week before I'm able to) and I don't have access to a Windows box (skimming
over, it looks like installation requires Windows). If it can be used on
Mac/Linux and the delay's not a problem, I'd be happy to review though.
