### 2018-07-09 #804 @arfon
OK thanks @bovee.

> If it can be used on Mac/Linux and the delay's not a problem, I'd be happy to review though.

@chasemc - can this software be used on Mac/Linux?
