### 2018-07-09 #804 @chasemc
Unfortunately not, I was waiting to address this as it was also a comment here (https://github.com/openjournals/joss-reviews/issues/804#issuecomment-403182850).

The reliance on MSConvert to convert spectra is necessary, and MSConvert itself relies on vendor dll files that require a Windows operating system.  There are workarounds for Linux and Mac, but they remain buggy and incomplete (https://sourceforge.net/p/proteowizard/mailman/message/35780564).
