### 2018-07-26 #804 @arfon
@chasemc - could you take a look at this list of viewers and see if you can identify any other potential candidates? https://bit.ly/joss-reviewers
