### 2018-07-30 #804 @mwang87
@chasemc Would be happy to review. I have my personal windows machine at home that I can install it on.

--Ming
