### 2018-07-31 #804 @lgatto
@chasemc - I don't have any windows computer at hand for the moment, and won't before a couple of months.
What about @jotsetung? Or @vladpetyuk? Do you have access to a windows.
