### 2018-08-03 #804 @chasemc
@arfon How many reviewers does JOSS usually have? I can see if I can suggest others if needed.
