### 2018-08-06 #804 @jotsetung
I do have a Windows (7) virtual machine for testing purposes, but I'm still in post-vacation stress and quite busy the next weeks.
