### 2018-08-29 #804 @chasemc
I think mass spec is especially well covered.  I went through the google doc again and think the following people (in no specific order) would also be qualified for review.
@juanvillada
@road2stat
@skummerf
@mschubert
@semacu
