### 2018-09-01 #804 @arfon
@juanvillada, @road2stat, @skummerf, @mschubert, @semacu - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
