### 2018-09-01 #804 @mschubert
I don't have a Windows box. - Judging from the comments, this seems to be a requirement.
