### 2018-09-03 #804 @semacu
I am unable to grade the submission as I am not an expert on this area and I do not have access to a Windows machine to try / test mzEasy in detail. Here are some points though:
- It would be useful if the paper could mention any ongoing research projects ot recent publications using mzEasy specifically
- Contains MIT LICENSE file
- Statement of need: it would be useful if the author could spell out both in the paper and the documentation the meaning of the acronym GNPS
- Installation instructions: Good, some dependencies (RInno / DesktopDeplyR) seem to be automatically installed after running the exe installer. However others (Proteowizard) require separate installation. Can all the different dependencies be installed at once?
- No example usage presented in the README. Please include a step-by-step example of how to use  mzEasy in a real dataset.
- Contains guidelines in CONTRIBUTING.md
- I could not try / test mzEasy as I don't have access to Windows (In a separate note, it'd be good to include an extra column in the joss_reviewers excel sheet to add a list of accessible OS by each potential reviewer)
