### 2018-09-03 #804 @chasemc
Thank you for your time @semacu.  I've informally outlined some responses below.

- It would be useful if the paper could mention any ongoing research projects ot recent publications using mzEasy specifically
  - I don't personally have any current projects using it, but certainly will within the next 6 months and it would have been useful for past work.  It has been downloaded > 70 times in 10 countries and I personally know a few people using it, but no publication.
- Statement of need: it would be useful if the author could spell out both in the paper and the documentation the meaning of the acronym GNPS
  - Will do!
- Installation instructions: Good, some dependencies (RInno / DesktopDeplyR) seem to be automatically installed after running the exe installer. However others (Proteowizard) require separate installation. Can all the different dependencies be installed at once?
  - Unfortunately not for Proteowizard for the same reason the program requires Windows (https://github.com/openjournals/joss-reviews/issues/804#issuecomment-403336901)- the vendor-supplied dlls in Proteowizard are windows based and the vendor-licenses are not all open for secondary distribution.
- No example usage presented in the README. Please include a step-by-step example of how to use mzEasy in a real dataset.
  - Will do!

Thanks again for the extra effort.
-Chase
