### 2018-09-24 ce7766b
Edits from pre-review suggestions
Modifications to text including suggestions from:
https://github.com/openjournals/joss-reviews/issues/804#issuecomment-403182850

and
https://github.com/openjournals/joss-reviews/issues/804#issuecomment-418193001

Will try to get to the Readmre edits tomorrow,  as suggested in: https://github.com/openjournals/joss-reviews/issues/804#issuecomment-418193001
