+++ paper/paper.bib
@@ -61,7 +61,7 @@ url = {www.ficonsulting.com},
     year = {2018},
     url = {https://www.R-project.org/},
   }
-  
+
 @article{Pedrioli2004,
 abstract = {A common open representation of mass spectrometry data and its application to proteomics research},
 author = {Pedrioli, Patrick G A and Eng, Jimmy K and Hubley, Robert and Vogelzang, Mathijs and Deutsch, Eric W and Raught, Brian and Pratt, Brian and Nilsson, Erik and Angeletti, Ruth H and Apweiler, Rolf and Cheung, Kei and Costello, Catherine E and Hermjakob, Henning and Huang, Sequin and Julian, Randall K and Kapp, Eugene and McComb, Mark E and Oliver, Stephen G and Omenn, Gilbert and Paton, Norman W and Simpson, Richard and Smith, Richard and Taylor, Chris F and Zhu, Weimin and Aebersold, Ruedi},
@@ -78,3 +78,24 @@ url = {http://www.nature.com/articles/nbt1031},
 volume = {22},
 year = {2004}
 }
+
+@Article{Pluskal2010,
+author="Pluskal, Tom{\'a}{\v{s}}
+and Castillo, Sandra
+and Villar-Briones, Alejandro
+and Ore{\v{s}}i{\v{c}}, Matej",
+title="MZmine 2: Modular framework for processing, visualizing, and analyzing mass spectrometry-based molecular profile data",
+journal="BMC Bioinformatics",
+year="2010",
+month="Jul",
+day="23",
+volume="11",
+number="1",
+pages="395",
+abstract="Mass spectrometry (MS) coupled with online separation methods is commonly applied for differential and quantitative profiling of biological samples in metabolomic as well as proteomic research. Such approaches are used for systems biology, functional genomics, and biomarker discovery, among others. An ongoing challenge of these molecular profiling approaches, however, is the development of better data processing methods. Here we introduce a new generation of a popular open-source data processing toolbox, MZmine 2.",
+issn="1471-2105",
+doi="10.1186/1471-2105-11-395",
+url="https://doi.org/10.1186/1471-2105-11-395"
+}
+
+
