+++ paper/paper.md
@@ -1,43 +1,57 @@
 ---
-title: 'mzEasy'
+title: 'mzEasy: A simple interface for mass spectrometry data conversion and LC-MS/MS
+  overview plotting'
 tags:
-  - Mass Spectrometry
-  - LC-MS/MS
-  - Liquid Chromatography
-  - File Conversion
-  - Quality Check
-authors:
- - name: Chase M Clark
-   orcid: 0000-0001-6439-9397
-   affiliation: "1"
+- Mass Spectrometry
+- LC-MS/MS
+- Liquid Chromatography
+- File Conversion
+- Quality Check
 affiliations:
- - name: University of Illinois at Chicago
-   index: 1
-date: 02 July 2018
+- index: 1
+  name: Department of Medicinal Chemistry and Pharmacognosy, College of Pharmacy, University of Illinois at Chicago, Chicago, IL
+authors:
+- affiliation: '1'
+  name: Chase M Clark
+  orcid: 0000-0001-6439-9397
+date: "02 July 2018"
+output:
+  pdf_document: default
+  html_document:
+    df_print: paged
 bibliography: paper.bib
+
 ---
 
 ### Summary
 
-An increase in access to liquid chromatography-mass spectrometry (LC-MS/MS) and advanced algorithms (often cloud-based) has led to a surge in users whose time is often better spent with advanced analyses than converting and evaluating, holistically, their raw data.  However, conversion of raw spectra to open access formats and ensuring data is of good quality are necessary steps for proper use in later analyses. Unfortunately, users unfamiliar with conversion software often miss needed settings or are unsure how to proceeed, adding a barrier of entry to the use of more advanced analysis pipelines.
+An increase in access to liquid chromatography-mass spectrometry (LC-MS/MS) and advanced algorithms (often cloud-based) has led to a surge in users whose time is often better spent with advanced analyses than converting and evaluating their raw data.  However, conversion of raw spectra to open access formats and ensuring data is of good quality are necessary steps for proper use in later analyses. Unfortunately, users unfamiliar with conversion software often miss needed settings or are unsure how to proceeed, adding a barrier of entry to the use of more advanced analysis pipelines.
 
-mzEasy provides a simple interface with MSConvert [@Chambers2012] to convert raw MS files to the "mzXML" open access format [@Pedrioli2004] and auto-generate quality control graphs (with an option to save graphs as a pre-assembled manuscript image). Importantly, mzEasy doesn't require (or have options for) adjusting/selecting any settings other than what image format to save plots as.  
+mzEasy provides a simple interface with MSConvert [@Chambers2012] to convert raw MS files to the "mzXML" open access format [@Pedrioli2004]. As it utilizes Proteowizard's MSConvert, it is capable of converting raw data from such vendors as AB SCIEX, Agilent, Bruker, and ThermoFisher, and some limited Waters' data. The auto-generated overview plots allow for easy inspection of the converted files and feedback of whether the conversion was succesful (including the option to save graphs as a pre-assembled manuscript image). Importantly, mzEasy doesn't require (or have options for) adjusting/selecting any settings other than to select which image format to save plots. For more advanced processing options another software such as MZmine 2 [@Pluskal2010] should be used.
 
-mzEasy is auto-configured to convert vendor files to mzXML with settings required for analyses such as GNPS [@Wang2016] and is meant for anyone uneasy with using MSConvertGUI.  It is also designed for users that want access to parallelized MSConvert and/or a simple way to generate quality control plots that are pre-arranged for easy publication/sharing.
[…]
