+++ paper/paper.md
@@ -62,10 +62,10 @@ It is also designed for users that want access to parallelized MSConvert and/or 
 mzEasy was written in R [@baseR2018] and utilizes Shiny [@shiny2018], RInno [@RInno2018], mzR [@Chambers2012], and ggplot2 [@gg2009].
 
 
-![A screenschot of the mzEasy user interface](Capture.PNG)
+![A screenschot of the mzEasy user interface](Capture.PNG){ width=75% }
 
 
-![Example of an exported, auto-assembled figure](Output.png)
+![Example of an exported, auto-assembled figure](Output.png){ width=75% }
 
 # References
 
