### 2018-09-24 #928 @chasemc
Images were out of place, needed to play with markdown
