### 2018-11-04 #928 @biswapriyamisra
Hello,

This is all going bit out of hand. The situation is funny and precarious
(with so many Emails, and so many invitees from  the List of Reviewers
declining for what ever reasons!), and then am being asked for the comments
again and again. I sympathize with the author- for this is all sounding
painful for a small software with 1-2 pages of content for the manuscript.
I hope in the future you would take care of these incomprehensible
situations, Emails, declining reviewers etc.

I did provide the 1st comments and an annotated PDF for the article in July
6th here https://github.com/openjournals/joss-reviews/issues/804 . These
are as follows and I have also attached the annotated PDF yet again for the
author.

   1. *Fairly quick and easy installation when MSConvert and R 3.5 or so
   are in place!*
   2. *Useful too for the conversions.*
   3. *More comments and suggestions in the annotated PDF.*
   4. *Useful tutorial and walk through images.*

*mzEasy_JOSS_Reviewer_Comments.pdf
<https://github.com/openjournals/joss-reviews/files/2172385/mzEasy_JOSS_Reviewer_Comments.pdf>*

*Thanks for the opportunity to review and all the best. Shall keep adding
requests to more functionalities or issues when faced.*

*Thanks, Biswa*

Thanks again for the opportunity to review it!

Best wishes,
Biswa
