### 2018-11-04 #928 @arfon
> I did provide the 1st comments and an annotated PDF for the article in July 6th here #804 . These are as follows and I have also attached the annotated PDF yet again for the author.

Hi @biswapriyamisra - sorry for causing frustration here but it's not clear from your comments on this PDF that you have actually carried out a review of this software. JOSS reviews are primarily *not* of the paper, rather the software instead.

As you can see at the top of this issue, there is a checklist for you to work through which _is the JOSS review_. The comments you have left in your annotated PDF are helpful but not sufficient for the level of review we ask of our reviewers.

If you are still willing, I'd ask that you familiarize yourself with the [reviewer guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) and work through the checklist above, leaving comments either here or on the review repository (https://github.com/chasemc/mzEasy).

Thanks again. Arfon.
