### 2019-01-13 #928 @arfon
On closer inspection of this submission I've deemed it not suitable for JOSS due to the minor utility nature of this package.

Apologies for not catching this sooner @chasemc.
